Introduce OpenACC

+++ packages/ml/src/Operator/ml_operator.h
@@ -299,7 +299,7 @@ extern int ML_Operator_Check_Getrow(ML_Operator *, int, char*);
 extern double ML_Operator_MaxNorm(ML_Operator *matrix, int divide_diag);
 extern double ML_Operator_FroNorm(ML_Operator *matrix, int divide_diag);
 extern int ML_Operator_Print(ML_Operator *matrix, const char label[]);
-extern int ML_Operator_ComputeNumNzs(ML_Operator *matrix);
+extern long long ML_Operator_ComputeNumNzs(ML_Operator *matrix);
 /* Operator Scaling */
 extern int ML_implicitscale_Getrow(ML_Operator *data, int N_requested_rows,
 				   int requested_rows[], int allocated_space,
